requirements.txt
groq==0.26.0
python-dotenv==1.0.0